+++ lib/online_migrations/background_migrations/migration_helpers.rb
@@ -395,6 +395,44 @@ module OnlineMigrations
       end
       alias remove_background_migration remove_background_data_migration
 
+      # Ensures that the background data migration with the provided configuration succeeded.
+      #
+      # If the enqueued migration was not found in development (probably when resetting a dev environment
+      # followed by `db:migrate`), then a log warning is printed.
+      # If enqueued migration was not found in production, then the error is raised.
+      # If enqueued migration was found but is not succeeded, then the error is raised.
+      #
+      # @param migration_name [String, Class] Background migration job class name
+      # @param arguments [Array, nil] Arguments with which background migration was enqueued
+      #
+      # @example Without arguments
+      #   ensure_background_data_migration_succeeded("BackfillProjectIssuesCount")
+      #
+      # @example With arguments
+      #   ensure_background_data_migration_succeeded("CopyColumn", arguments: ["users", "id", "id_for_type_change"])
+      #
+      def ensure_background_data_migration_succeeded(migration_name, arguments: nil)
+        migration_name = migration_name.name if migration_name.is_a?(Class)
+
+        configuration = { migration_name: migration_name }
+
+        if arguments
+          arguments = Array(arguments)
+          migration = Migration.parents.for_configuration(migration_name, arguments).first
+          configuration[:arguments] = arguments.to_json
+        else
+          migration = Migration.parents.for_migration_name(migration_name).first
+        end
+
+        if migration.nil?
+          Utils.raise_in_prod_or_say_in_dev("Could not find background data migration for the given configuration: #{configuration}")
+        elsif !migration.succeeded?
+          raise "Expected background data migration for the given configuration to be marked as 'succeeded', " \
+                "but it is '#{migration.status}': #{configuration}"
+        end
+      end
+      alias ensure_background_migration_succeeded ensure_background_data_migration_succeeded
+
       # @private
       def create_background_data_migration(migration_name, *arguments, **options)
         options.assert_valid_keys(:batch_column_name, :min_value, :max_value, :batch_size, :sub_batch_size,